requirements.txt
# LangChain
requests
beautifulsoup4
pdfplumber
tqdm
langchain
langchain-community
langchain-openai
langchain-mistralai

# Vector store
faiss-cpu

# LLM / Embeddings
openai

# Document loaders
pdfplumber
python-docx

# Config & env
python-dotenv
pydantic
pyyaml

# Web search
tavily-python
ddgs

# Interface
chainlit
streamlit

# Tests
pytest

# API (optionnel)
fastapi
uvicorn